refactoring of ExplorerDetails and websocket subscription for token history
